+++ examples/demo-rollup/src/lib.rs
@@ -4,6 +4,8 @@
 // TODO: #![doc = include_str!("../README.md")]
 #![deny(missing_docs)]
 
+use std::str::FromStr;
+
 use const_rollup_config::{ROLLUP_BATCH_NAMESPACE_RAW, ROLLUP_PROOF_NAMESPACE_RAW};
 use sov_celestia_adapter::types::Namespace;
 
@@ -15,8 +17,6 @@ mod celestia_rollup;
 
 pub use celestia_rollup::*;
 
-mod eth;
-
 mod zk;
 pub use zk::*;
 
@@ -26,3 +26,11 @@ pub const ROLLUP_BATCH_NAMESPACE: Namespace = Namespace::const_v0(ROLLUP_BATCH_N
 
 /// The rollup stores the zk proofs in the namespace b"sov-test-p" on Celestia.
 pub const ROLLUP_PROOF_NAMESPACE: Namespace = Namespace::const_v0(ROLLUP_PROOF_NAMESPACE_RAW);
+
+// TODO: https://github.com/Sovereign-Labs/sovereign-sdk-wip/issues/387
+fn eth_dev_signer() -> sov_ethereum::DevSigner {
+    sov_ethereum::DevSigner::new(vec![secp256k1::SecretKey::from_str(
+        "ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80",
+    )
+    .unwrap()])
+}