+++ examples/demo-rollup/tests/evm/mod.rs
@@ -1,4 +1,3 @@
 mod evm_account_abstraction;
 mod evm_test_helper;
 mod evm_tx;
-mod test_client;